refactor(waveform-seek): H3.6 — extract useWaveformInterpolation hook
